requirements.txt
anyio==4.2.0
asn1==2.7.0
async-timeout==4.0.3
base58==2.1.1
bleak==0.21.1
bleak-winrt==1.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
ECPy==1.2.5
enum-compat==0.0.3
exceptiongroup==1.2.0
future==0.18.3
h11==0.14.0
hidapi==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
ledgerblue==0.1.48
libusb1==3.1.0
ndeflib==0.3.3
nfcpy==1.0.4
pillow==10.2.0
protobuf==4.25.1
pycryptodome==3.19.1
pycryptodomex==3.19.1
pyDes==2.0.1
pyelftools==0.30
pyeoskit==1.1.12
pyserial==3.5
python-dotenv==1.0.0
python-gnupg==0.5.2
python-u2flib-host==3.0.3
requests==2.31.0
requests-unixsocket==0.3.0
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
websocket-client==1.7.0
